Combine helpers-declarative and helpers-results.
